backend/requirements.txt
asgiref==3.4.1
astroid==2.9.3
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0
django-cors-headers==3.10.1
django-filter==21.1
django-stubs==1.9.0
django-stubs-ext==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
greenlet==1.1.2
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.931
mypy-extensions==0.4.3
numpy==1.22.0
oauthlib==3.1.1
packaging==21.3
pandas==1.3.5
platformdirs==2.4.1
psycopg2-binary==2.9.3
py4j==0.10.9.2
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
pylint-plugin-utils==0.7
pyparsing==3.0.6
pyspark==3.2.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.19
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
SQLAlchemy==1.4.29
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.0
types-pytz==2021.3.4
types-PyYAML==6.0.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3
